Use stock validator in the example
