+++ src/validators/regexp.js
@@ -1,9 +1,9 @@
-import isNil from '../helpers/isNil';
-
 /*eslint func-style:0*/
-export default function required(errorMessage) {
+export default function regexp(regexp, errorMessage) {
     return function (value) {
-        if (isNil(value) || !value.length) {
+        const r = regexp && regexp.test ? regexp : new RegExp(regexp);
+
+        if (r.test(value)) {
             return errorMessage;
         }
     };
